// typescript/user-interface/packages/weavess/src/ts/components/waveform-display/types.ts
export interface WaveformDisplayProps {
  /** waveform display configuration */
  initialConfiguration?: Partial<WeavessTypes.Configuration>;

  /** flex or not? */
  flex?: boolean;

  /** viewable time interval, full amount of data window */
  viewableInterval: WeavessTypes.TimeRange;

  /** The minimum offset in seconds */
  minimumOffset: number;

  /** The maximum offset in seconds */
  maximumOffset: number;

  /** callback executed when closing expanded mode */
  closeSignalDetectionOverlayCallback?: () => void;

  /** The arrival time of the station nearest the open event */
  baseStationTime?: number;

  /** stations */
  stations: WeavessTypes.Station[];

  /** events */
  events: WeavessTypes.Events;

  /** Is the main waveform panel controlled or uncontrolled */
  /** default assume controlled */
  isControlledComponent?: boolean;

  /**
   * The current interval might be different than viewable interval.
   * Will default to use viewableInterval if not set.
   */
  currentInterval?: WeavessTypes.TimeRange;

  /** selections */
  selections?: WeavessTypes.Selections;

  /** markers */
  markers?: WeavessTypes.Markers;

  /** Defines a custom component for displaying a custom label */
  customLabel?: React.FunctionComponent<WeavessTypes.LabelProps>;

  /** display the measure window */
  showMeasureWindow?: boolean;

  /** specifies the measure window selection */
  measureWindowSelection?: WeavessTypes.MeasureWindowSelection;

  uiTheme?: {
    isLightMode: boolean;
  };

  unassociatedSDColor?: string;

  /** The amount to zoom in when using the zoom hotkeys. Defaults to 0.5. */
  hotkeyZoomInRatio?: number;

  /** The amount to zoom out when using the zoom hotkeys. Defaults to 0.6666666667. */
  hotkeyZoomOutRatio?: number;

  /** The ratio of the screen to pan when using hotkeys, where 0 is none, and 1 is the full screen. Defaults to 0.25 */
  panRatio?: number;

  /**
   * By default, WEAVESS creates its own toast container. If one is provided by the parent, set this to true in order
   * avoid having multiple toasts.
   */
  disableToastContainer?: boolean;

  /** a JSX element that will render to the right of the time axis to display extra info */
  extraInfoBar?: JSX.Element;

  /** event handler for clearing selected channels */
  clearSelectedChannels?(): void;

  /**
   * Event handler for selecting a channel
   *
   * @param channelId a Channel Id as a string
   */
  selectChannel?(channelId: string): void;

  /**
   * Used to look up the position buffer data (a Float32Array formatted like so: x y x y x y ...).
   * Takes the position buffer's id.
   */
  getPositionBuffer?(
    id: string,
    startTime: number,
    endTime: number,
    domainTimeRange: WeavessTypes.TimeRange
  ): Promise<Float32Array>;

  /**
   * Used to look up the Channel Segment Boundaries for a given channel segment by name
   */
  getBoundaries?(
    channelName: string,
    channelSegment: WeavessTypes.ChannelSegment,
    timeRange?: WeavessTypes.TimeRange,
    isMeasureWindow?: boolean
  ): Promise<WeavessTypes.ChannelSegmentBoundaries>;

  onMount?(weavessInstance: WeavessTypes.WeavessInstance): void;

  setViewportVisibleStations?(
    channels: WeavessTypes.Channel[],
    indexStart: number,
    indexEnd: number
  ): void;
}

export interface WaveformDisplayState {
  /** Configuration for waveform display */
  initialConfiguration: WeavessTypes.Configuration;

  /** display the measure window */
  showMeasureWindow?: boolean;

  /** Determines if the measure window is displayed */
  isMeasureWindowVisible: boolean;

  /** Height of the measure window in pixels */
  measureWindowHeightPx: number;

  /** Selection info needed to render a measure window */
  measureWindowSelection: WeavessTypes.MeasureWindowSelection | undefined;

  /**
   * the previous measure window selection passed in from props
   * (used to ensure the measure window is only updated when expected)
   */
  prevMeasureWindowSelectionFromProps: WeavessTypes.MeasureWindowSelection | undefined;

  /** true if waveforms should be rendered; false otherwise */
  shouldRenderWaveforms: boolean;

  /** true if spectrograms should be rendered; false otherwise */
  shouldRenderSpectrograms: boolean;
}